Fixed big drawback of `@typefreeze` approach: dont add methods to user function
Before this commit, the biggest problem with the approach of memoizing
using dispatch was that by adding methods to the user's generic function
object, we were adding a constraint, that you could not redefine the
`@typefreeze`d method, since the more-specific generated methods would
always be preferred.

After this commit, those methods are hidden on a gensym'd generic
function, and not not exposed to users, so that redefining the
`@typefreeze`d method simply delegates to a new helper method, and all
previously generated methods can be forgotten (and possibly GC'd).

The other downside of increasing the cost for dynamic dispatch by
increasing the number of methods to search still remains.